requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
Pillow==10.2.0
python-decouple==3.8
celery==5.3.6
redis==5.0.1
stripe==8.1.0
pandas==2.2.0
numpy==1.26.3
django-filter==23.5
drf-spectacular==0.27.0
